+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 **Fixes:**
 - **OpenShift:** Fixed issue where the Trident Operator fails to patch ClusterRole and ClusterRoleBinding (Issue [#517](https://github.com/NetApp/trident/issues/517)).
 - Fixed a parsing error when iscsiadm listed target portals with a negative group tag
+- Fixed issue where docker plugin could not be upgraded from v19.10 to v21.01 (Issue [#507](https://github.com/NetApp/trident/issues/507))
 
 **Enhancements:**
 - Updated to golang 1.15